Simplify project and platform build boundaries
