+++ crossbundle/cli/Cargo.toml
@@ -32,6 +32,7 @@ pretty_env_logger = { workspace = true }
 log = { workspace = true }
 
 dunce = { workspace = true }
+home = { workspace = true }
 ureq = { workspace = true }
 
 [dev-dependencies]
