+++ crossbundle/tools/Cargo.toml
@@ -73,6 +73,7 @@ toml = { workspace = true }
 xml-rs = { workspace = true, optional = true }
 
 dunce = { workspace = true }
+home = { workspace = true }
 zip = { workspace = true, default-features = false, features = ["deflate-flate2-zlib-rs"] }
 image = { workspace = true, default-features = false, features = ["png", "jpeg"] }
