+++ crossbundle/tools/src/commands/android/common/gen_key.rs
@@ -8,50 +8,33 @@ pub fn gen_key(
     sign_key_pass: Option<String>,
     sign_key_alias: Option<String>,
 ) -> crate::error::Result<Key> {
-    let key = if let Some(key_path) = sign_key_path {
-        let aab_key = Key {
+    let key = match (sign_key_path, sign_key_pass, sign_key_alias) {
+        (Some(key_path), Some(key_pass), Some(key_alias)) => Key {
             key_path,
-            key_pass: sign_key_pass.unwrap(),
-            key_alias: sign_key_alias.unwrap(),
-        };
-        if aab_key.key_path.exists() {
-            aab_key
-        } else {
-            Keytool::new()
-                .genkeypair(true)
-                .v(true)
-                .keystore(&aab_key.key_path)
-                .alias(&aab_key.key_alias)
-                .keypass(&aab_key.key_pass)
-                .storepass(&aab_key.key_pass)
-                .dname(&["CN=Android Debug,O=Android,C=US".to_owned()])
-                .keyalg(KeyAlgorithm::RSA)
-                .keysize(2048)
-                .validity(10000)
-                .run()?
-                // This will never panic because of AabKey always returned if help flag not set
-                .unwrap()
+            key_pass,
+            key_alias,
+        },
+        (Some(_), _, _) => {
+            return Err(
+                anyhow::anyhow!("a signing key requires both a password and an alias").into(),
+            );
         }
-    } else {
-        let aab_key = Key::new_default()?;
-        if aab_key.key_path.exists() {
-            aab_key
-        } else {
-            Keytool::new()
-                .genkeypair(true)
-                .v(true)
-                .keystore(&aab_key.key_path)
-                .alias(&aab_key.key_alias)
-                .keypass(&aab_key.key_pass)
-                .storepass(&aab_key.key_pass)
-                .dname(&["CN=Android Debug,O=Android,C=US".to_owned()])
-                .keyalg(KeyAlgorithm::RSA)
-                .keysize(2048)
-                .validity(10000)
-                .run()?
-                // This will never panic because of AabKey always returned if help flag not set
-                .unwrap()
-        }
+        (None, _, _) => Key::new_default()?,
     };
-    Ok(key)
+    if key.key_path.exists() {
+        return Ok(key);
+    }
+    Keytool::new()
+        .genkeypair(true)
+        .v(true)
+        .keystore(&key.key_path)
+        .alias(&key.key_alias)
+        .keypass(&key.key_pass)
+        .storepass(&key.key_pass)
+        .dname(&["CN=Android Debug,O=Android,C=US".to_owned()])
+        .keyalg(KeyAlgorithm::RSA)
+        .keysize(2048)
+        .validity(10000)
+        .run()?
+        .ok_or_else(|| anyhow::anyhow!("keytool did not return the generated signing key").into())
 }
