+++ crossbundle/tools/src/commands/android/common/helper_functions.rs
@@ -2,13 +2,12 @@ use crate::error::*;
 
 /// Helper function to delete files
 pub fn remove(target: Vec<std::path::PathBuf>) -> Result<()> {
-    target.iter().for_each(|content| {
-        if content.exists() && content.is_file() {
-            std::fs::remove_file(content).unwrap();
+    for content in target {
+        if content.is_file() {
+            std::fs::remove_file(content)?;
+        } else if content.is_dir() {
+            std::fs::remove_dir_all(content)?;
         }
-        if content.exists() && content.is_dir() {
-            std::fs::remove_dir_all(content).unwrap();
-        }
-    });
+    }
     Ok(())
 }
