+++ crossbundle/tools/src/types/common/mod.rs
@@ -1,12 +1,12 @@
 mod build_variables;
-mod config;
+mod cli_context;
 mod profile;
 mod shell;
 mod target;
 mod version;
 
 pub use build_variables::*;
-pub use config::*;
+pub use cli_context::*;
 pub use profile::*;
 pub use shell::*;
 pub use target::*;
